reguster route deleted and merged into the auth.py route

+++ Flask/todo_app/app/routes/tasks.py
@@ -12,7 +12,7 @@ def view_tasks() :
     if 'user' not in session :
         return redirect(url_for('auth.login'))
     
-    tasks = Task.query.all()
+    tasks = Task.query.filter_by(user_id = session['user']).all()
     return render_template('tasks.html', tasks = tasks)
 
 
@@ -41,7 +41,7 @@ def add_task() :
 @tasks_bp.route('/toggle/<int:task_id>', methods = ['POST'])
 def toggle_status(task_id) : 
 
-    task = Task.query.filter_by(user_id = session['user']).all()
+    task = Task.query.filter_by(id = task_id, user_id = session['user']).first()
 
     if task  :
 
@@ -59,7 +59,7 @@ def toggle_status(task_id) :
 @tasks_bp.route('/clear', methods = ['POST'])
 def clear_tasks() : 
 
-    Task.query.delete()
+    Task.query.filter_by(user_id = session['user']).delete()
     db.session.commit()
     flash('All tasks cleared', 'info')
